feat: switched to mise version manager

+++ zsh/.zsh/alias.zsh
@@ -1,17 +1,14 @@
 # GLOBAL--------------------------------------------------------------------------------------------
 
+## vim
+alias vim="nvim"
+
 ## Replaces `ls` command with `eza`
 alias ls="eza --icons --group-directories-first"
 alias ll="eza --icons --group-directories-first -l"
 
 ## cat
-if [[ "$OSTYPE" == "darwin"* ]]; then
-  alias cat='bat --decorations=never'
-fi
-if [[ "$OSTYPE" == "linux-gnu"* ]]; then
-  alias cat='batcat --decorations=never'
-  alias bat='batcat'
-fi
+alias cat='bat --decorations=never'
 
 ## df and du
 alias df="df -hP"
@@ -28,28 +25,33 @@ alias md='glow'
 alias t="tmux new -s '󱂶 HOME'"
 alias ta="tmux a"
 
+## PATH
+alias path='echo $path | tr " " "\n" | bat --file-name PATH'
 
 # MAC OS--------------------------------------------------------------------------------------------
+if [[ "$OSTYPE" == "darwin"* ]]; then
 
-## Flush system dns
-alias flushhostdns="sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder"
+    ## Flush system dns
+    alias flushhostdns="sudo dscacheutil -flushcache; sudo killall -HUP mDNSResponder"
 
-## Recursively delete `.DS_Store` files
-alias cleanup="find . -type f -name '*.DS_Store' -ls -delete"
+    ## Recursively delete `.DS_Store` files
+    alias cleanup="find . -type f -name '*.DS_Store' -ls -delete"
 
-## Plist Buddy
-alias plistbuddy="/usr/libexec/PlistBuddy"
+    ## Plist Buddy
+    alias plistbuddy="/usr/libexec/PlistBuddy"
 
-## Hexyl (hex viewer)
-alias hex="hexyl --terminal-width 150"
+    ## Hexyl (hex viewer)
+    alias hex="hexyl --terminal-width 150"
 
-## LeetCode (leetcode.nvim)
-alias leetcode="nvim leetcode.nvim"
+fi
 
 # SERVER--------------------------------------------------------------------------------------------
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
 
-## Docker alias
-alias dps="docker ps --format 'table {{.Names}}\t {{.ID}}\t {{.Image}}\t {{.Status}}\t'"
-alias dcup="docker compose up -d"
-alias dcupr="docker compose up -d --force-recreate"
-alias dcl="docker compose logs -f"
+    ## Docker alias
+    alias dps="docker ps --format 'table {{.Names}}\t {{.ID}}\t {{.Image}}\t {{.Status}}\t'"
+    alias dcup="docker compose up -d"
+    alias dcupr="docker compose up -d --force-recreate"
+    alias dcl="docker compose logs -f"
+
+fi
